park-finder-api/cronjob/requirements.txt
APScheduler==3.3.0 
autopep8==2.0.2
flask-cors==3.0.10
python-dotenv==1.0.0
pymongo==3.12.3
pymongo[srv]
flask==2.1.3
Werkzeug==2.3.7
requests